Refactor ArticleRestController with different RestController
