+++ src/main/kotlin/io/github/raeperd/realworldspringbootkotlin/web/ArticleRestController.kt
@@ -43,18 +43,6 @@ class ArticleRestController(
             .toMultipleArticleModel()
     }
 
-    @PostMapping("/articles/{slug}/favorite")
-    fun postArticleFavoriteBySlug(@PathVariable slug: String, payload: JWTPayload): ArticleModel {
-        return articleService.favoriteArticle(payload.sub, slug).toArticleModel()
-    }
-
-    @DeleteMapping("/articles/{slug}/favorite")
-    fun deleteArticleFavoriteBySlug(@PathVariable slug: String, payload: JWTPayload): ArticleModel {
-        return articleService.unfavoriteArticle(payload.sub, slug).toArticleModel()
-    }
-
-    private fun ArticleDTO.toArticleModel(): ArticleModel = ArticleModel(this)
-
     private fun Page<ArticleDTO>.toMultipleArticleModel() = MultipleArticleModel(content, content.size)
 }
 
@@ -95,3 +83,5 @@ data class ArticlePutDTO(
 data class ArticleModel(val article: ArticleDTO)
 
 data class MultipleArticleModel(val articles: List<ArticleDTO>, val articlesCount: Int)
+
+fun ArticleDTO.toArticleModel(): ArticleModel = ArticleModel(this)
